🔨 인증 상태와 프로필 가져오는 로직 분리
